Fix logrotate race conditions
