+++ tx-rx/conf/dxwifi.conf
@@ -1,7 +1,12 @@
 # DxWiFi Log rotation strategy - Place in /etc/logrotate.d/
+# Note, depending on the verbosity used when running dxwifi, you may have to 
+# update the logrotate cron job to hourly
+
 /var/log/dxwifi.log {
     rotate 6
     size 500k
     compress
+    delaycompress
+    create
     missingok
 }
